Recreate empty folder, needed for testcases

+++ test/de/uni_hildesheim/sse/kernel_haven/kconfigreader/KconfigReaderExtractorTest.java
@@ -36,7 +36,9 @@ public class KconfigReaderExtractorTest {
     @After
     public void tearDown() {
         for (File file : RESOURCE_DIR.listFiles()) {
-            file.delete();
+            if (!file.getName().equals(".gitignore")) {
+                file.delete();
+            }
         }
     }
